feat: dockerized service.

+++ SweetManagerWebService/SupplyManagement/Domain/Model/Commands/CreateSuppliesRequestCommand.cs
@@ -1,3 +1,3 @@
 ﻿namespace SweetManagerWebService.SupplyManagement.Domain.Model.Commands;
 
-public record CreateSuppliesRequestCommand(int PaymentsOwnersId, int SuppliesId, int Count, int Amount); 
+public record CreateSuppliesRequestCommand(int PaymentsOwnersId, int SuppliesId, int Count, decimal Amount);
